api/requirements.txt
attrs==18.1.0
click==6.7
Flask==1.0
Flask-MySQL==1.4.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.1.0
mysql-connector==2.1.6
pluggy==0.6.0
py==1.5.3
PyMySQL==0.8.1
pytest==3.5.1
six==1.11.0
Werkzeug==0.14.1
